requirements.txt
Flask==0.12.2
coinbase==2.0.6
gunicorn==19.9.0
python-binance==0.5.15
